Cambio en la implementación de NodoWritable Modificación de las clases Mapper y Reducer para trabajar con SequenceFiles

+++ WikipediaPageRank/src/icesi/ugoogle/pagerank/WikipediaFinalPageRankReducer.java
@@ -2,25 +2,25 @@ package icesi.ugoogle.pagerank;
 
 import java.io.IOException;
 
-import org.apache.hadoop.io.FloatWritable;
+import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.ObjectWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Reducer;
 
 public class WikipediaFinalPageRankReducer extends
-		Reducer<Text, ObjectWritable, Text, FloatWritable> {
+		Reducer<Text, ObjectWritable, Text, DoubleWritable> {
 
 	protected void reduce(Text word, Iterable<ObjectWritable> values,
 			Context context) throws IOException, InterruptedException {
-		float pr = 0;
+		double pr = 0;
 
 		for (ObjectWritable w : values) {
 			if (w.getDeclaredClass().toString()
-					.equals(FloatWritable.class.toString())) {
-				pr += ((FloatWritable) w.get()).get();
+					.equals(DoubleWritable.class.toString())) {
+				pr += ((DoubleWritable) w.get()).get();
 			}
 		}
-		FloatWritable value = new FloatWritable(pr);
+		DoubleWritable value = new DoubleWritable(pr);
 		context.write(word, value);
 	}
 }
